back/requirements.txt
# ============================================
# requirements.txt - LIMPIO Y ORGANIZADO
# ============================================

# Core FastAPI Stack
fastapi>=0.115.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary
python-dotenv==1.0.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
authlib==1.2.1
itsdangerous==2.1.2

# HTTP & Networking
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6
urllib3<2.0

# Configuration & Utils
PyYAML>=6.0
click>=8.0.0

# Development Tools (solo si necesitas)
black==24.4.2
isort==5.12.0
flake8==6.1.0
mypy==1.10.0
pre-commit==3.7.0

# Testing (solo si necesitas)
pytest==7.4.3
pytest-asyncio==0.21.1
